test_compute_shader
